Refactor frontend to match backend changes (#5)

+++ server/src/domain/cloud_rider_message.rs
@@ -2,7 +2,7 @@ use chrono::Utc;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload")]
+#[serde(tag = "type", content = "data")]
 pub enum CloudRiderMessage {
     Heartbeat(Heartbeat),
     GlobalPosition(GlobalPosition),
@@ -12,10 +12,10 @@ pub enum CloudRiderMessage {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct GlobalPosition {
-    pub lat: i32,
-    pub lon: i32,
-    pub alt: i32,
-    pub relative_alt: i32,
+    pub lat: f64,
+    pub lon: f64,
+    pub alt: f64,
+    pub relative_alt: f64,
     pub vx: i16,
     pub vy: i16,
     pub vz: i16,
@@ -37,6 +37,6 @@ impl Heartbeat {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BatteryStatus {
     pub remaining_percent: i8,
-    pub temperature_c: i16,
+    pub temperature_c: f32,
     pub current_battery_ma: i16,
 }
